+++ test/test_helper.exs
@@ -5,4 +5,11 @@ exclude =
     []
   end
 
+exclude =
+  if System.get_env("MYSQL_URL") in [nil, ""] do
+    [:mysql | exclude]
+  else
+    exclude
+  end
+
 ExUnit.start(exclude: exclude)